# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlashHunt API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /friends/search:
    get:
      operationId: searchUsers
      tags: [friends]
      summary: Search users by username or email
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /friends/request:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFriendRequestBody"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/request/{requestId}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
  /friends/request/{requestId}/reject:
    post:
      operationId: rejectFriendRequest
      tags: [friends]
      summary: Reject friend request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Request rejected
  /friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Friend removed
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        totalPoints:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - username
        - totalPoints
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        city:
          type: string
        avatarUrl:
          type: string
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        totalPoints:
          type: integer
      required:
        - id
        - username
        - totalPoints
...
    Friend:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        lastActivity:
          type: ["string", "null"]
        lastItemFound:
          type: ["string", "null"]
        weekPoints:
          type: integer
      required:
        - id
        - userId
        - username
        - weekPoints
    FriendRequest:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: string
        toUserId:
          type: string
        fromUsername:
          type: string
        fromDisplayName:
          type: ["string", "null"]
        fromAvatarUrl:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - fromUserId
        - toUserId
        - fromUsername
        - status
        - createdAt
    SendFriendRequestBody:
      type: object
      properties:
        toUserId:
          type: string
      required:
        - toUserId
...